requirements.txt
praw>=7.7.1
anthropic>=0.25.0
flask>=3.0.0
python-dotenv>=1.0.0
